fix query to new schema, optimize query (filter first join later)

+++ queries/refund.sql
@@ -12,18 +12,29 @@ SELECT EXISTS (
 ) AS exists;
 
 -- name: GetRefund :one
-SELECT 
-  r.*,
-  COALESCE(array_agg(DISTINCT res.url) FILTER (WHERE res.url IS NOT NULL), '{}')::text[] as resources
-FROM payment.refund r
-LEFT JOIN product.resource res ON r.id = res.owner_id
-INNER JOIN payment.product_on_payment pop ON r.product_on_payment_id = pop.id
-INNER JOIN payment.base p ON pop.payment_id = p.id
-WHERE (
-  r.id = $1 AND 
-  (p.user_id = sqlc.narg('user_id') OR sqlc.narg('user_id') IS NULL)
+WITH filtered_refund AS (
+    SELECT r.*
+    FROM payment.refund r
+    INNER JOIN payment.product_on_payment pop ON r.product_on_payment_id = pop.id
+    INNER JOIN payment.base p ON pop.payment_id = p.id
+    WHERE (
+        r.id = sqlc.arg('id') AND 
+        (p.user_id = sqlc.narg('user_id') OR sqlc.narg('user_id') IS NULL)
+    )
+),
+filtered_resource AS (
+    SELECT 
+        res.owner_id,
+        array_agg(res.url ORDER BY res.order ASC) AS resources
+    FROM product.resource res
+    WHERE res.url IS NOT NULL AND res.owner_id = sqlc.arg('id')
+    GROUP BY res.owner_id
 )
-GROUP BY r.id;
+SELECT 
+    r.*,
+    COALESCE(res.resources, '{}') AS resources
+FROM filtered_refund r
+LEFT JOIN filtered_resource res ON res.owner_id = r.id;
 
 -- name: CountRefunds :one
 SELECT COUNT(r.id)
@@ -42,51 +53,50 @@ WHERE (
 );
 
 -- name: ListRefunds :many
+WITH filtered_refund AS (
+    SELECT r.*
+    FROM payment.refund r
+    INNER JOIN payment.product_on_payment pop ON r.product_on_payment_id = pop.id
+    INNER JOIN payment.base p ON pop.payment_id = p.id
+    WHERE (
+        (p.user_id = sqlc.narg('user_id') OR sqlc.narg('user_id') IS NULL) AND
+        (r.product_on_payment_id = sqlc.narg('product_on_payment_id') OR sqlc.narg('product_on_payment_id') IS NULL) AND
+        (r.method = sqlc.narg('method') OR sqlc.narg('method') IS NULL) AND
+        (r.status = sqlc.narg('status') OR sqlc.narg('status') IS NULL) AND
+        (r.reason ILIKE '%' || sqlc.narg('reason') || '%' OR sqlc.narg('reason') IS NULL) AND
+        (r.address ILIKE '%' || sqlc.narg('address') || '%' OR sqlc.narg('address') IS NULL) AND
+        (r.date_created >= sqlc.narg('date_created_from') OR sqlc.narg('date_created_from') IS NULL) AND
+        (r.date_created <= sqlc.narg('date_created_to') OR sqlc.narg('date_created_to') IS NULL)
+    )
+),
+filtered_resource AS (
+    SELECT 
+        res.owner_id,
+        array_agg(res.url ORDER BY res.order ASC) AS resources
+    FROM product.resource res
+    WHERE res.url IS NOT NULL AND res.owner_id IN (SELECT id FROM filtered_refund)
+    GROUP BY res.owner_id
+)
 SELECT 
     r.*,
-    COALESCE(array_agg(DISTINCT res.url) FILTER (WHERE res.url IS NOT NULL), '{}')::text[] as resources
-FROM payment.refund r
-LEFT JOIN product.resource res ON res.owner_id = r.id
-INNER JOIN payment.product_on_payment pop ON r.product_on_payment_id = pop.id
-INNER JOIN payment.base p ON pop.payment_id = p.id
-WHERE (
-    (p.user_id = sqlc.narg('user_id') OR sqlc.narg('user_id') IS NULL) AND
-    (r.product_on_payment_id = sqlc.narg('product_on_payment_id') OR sqlc.narg('product_on_payment_id') IS NULL) AND
-    (r.method = sqlc.narg('method') OR sqlc.narg('method') IS NULL) AND
-    (r.status = sqlc.narg('status') OR sqlc.narg('status') IS NULL) AND
-    (r.reason ILIKE '%' || sqlc.narg('reason') || '%' OR sqlc.narg('reason') IS NULL) AND
-    (r.address ILIKE '%' || sqlc.narg('address') || '%' OR sqlc.narg('address') IS NULL) AND
-    (r.date_created >= sqlc.narg('date_created_from') OR sqlc.narg('date_created_from') IS NULL) AND
-    (r.date_created <= sqlc.narg('date_created_to') OR sqlc.narg('date_created_to') IS NULL)
-)
-GROUP BY r.id
+    COALESCE(res.resources, '{}') AS resources
+FROM filtered_refund r
+LEFT JOIN filtered_resource res ON res.owner_id = r.id
 ORDER BY r.date_created DESC
 LIMIT sqlc.arg('limit')
 OFFSET sqlc.arg('offset');
 
 -- name: CreateRefund :one
-WITH inserted_refund AS (
-    INSERT INTO payment.refund (
-        product_on_payment_id,
-        method,
-        status,
-        reason,
-        address
-    )
-    VALUES (
-        $1, $2, $3, $4, $5
-    )
-    RETURNING *
-),
-inserted_resources AS (
-    INSERT INTO product.resource (owner_id, url)
-    SELECT id, unnest(sqlc.arg('resources')::text[]) FROM inserted_refund
-    RETURNING url
+INSERT INTO payment.refund (
+    product_on_payment_id,
+    method,
+    status,
+    reason,
+    address
+) VALUES (
+    $1, $2, $3, $4, $5
 )
-SELECT r.id, COALESCE(array_agg(DISTINCT res.url) FILTER (WHERE res.url IS NOT NULL), '{}')::text[] as resources
-FROM inserted_refund r
-LEFT JOIN inserted_resources res ON true
-GROUP BY r.id;
+RETURNING *;
 
 -- name: UpdateRefund :exec
 UPDATE payment.refund r
@@ -106,7 +116,7 @@ WHERE (
 -- name: DeleteRefund :exec
 DELETE FROM payment.refund r
 WHERE r.id = $1
-  AND EXISTS (
+  AND EXISTS ( -- Check if the refund belongs to the user
     SELECT 1
     FROM payment.product_on_payment pop
     JOIN payment.base p ON pop.payment_id = p.id
